Refactor: Move stream model to the models.py module
